Split the file into functions and put them into separate files
