+++ homework-03/controllers/contacts/index.js
@@ -1,4 +1,18 @@
-const contacts = require('./contacts');
+const getAllContacts = require('./getAllContacts');
+const getContactById = require('./getContactById');
+const addContact = require('./addContact');
+const updateContactById = require('./updateContactById');
+const removeContactById = require('./removeContactById');
+const updateFavoriteContact = require('./updateFavoriteContact');
+
+const contacts = {
+	getAllContacts,
+	getContactById,
+	addContact,
+	updateContactById,
+	removeContactById,
+	updateFavoriteContact,
+};
 
 module.exports = {
 	contacts,
