+++ homework-03/controllers/contacts/updateContactById.js
@@ -1,0 +1,17 @@
+const { NotFound } = require('http-errors');
+
+const { sendSuccessRes } = require('../../utils');
+const { Contact } = require('../../models/contacts');
+
+const updateContactById = async (req, res, next) => {
+	const { id } = req.params;
+	const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
+		new: true,
+	});
+	if (!updatedContact) {
+		throw new NotFound(`Product with id=${id} not found`);
+	}
+	sendSuccessRes(res, { updatedContact });
+};
+
+module.exports = updateContactById;
